+++ src/main/java/Factory.java
@@ -24,14 +24,32 @@
  *
  */
 
-package com.chrisrm.ideaddlc.actions.indicators;
+import com.intellij.ide.todo.TodoView;
+import com.intellij.openapi.components.ServiceManager;
+import com.intellij.openapi.project.DumbService;
+import com.intellij.openapi.project.Project;
+import com.intellij.openapi.wm.ToolWindow;
+import com.intellij.openapi.wm.ToolWindowFactory;
+import org.jetbrains.annotations.NotNull;
 
-import com.chrisrm.ideaddlc.config.ui.IndicatorStyles;
-
-public final class MTDotIndicators extends MTAbstractIndicatorsAction {
+public final class Factory implements ToolWindowFactory {
+  @Override
+  public void createToolWindowContent(@NotNull final Project project, @NotNull final ToolWindow toolWindow) {
+    DumbService.getInstance(project).runWhenSmart(() -> ServiceManager.getService(project, TodoView.class).initToolWindow(toolWindow));
+  }
 
   @Override
-  protected IndicatorStyles getIndicatorStyle() {
-    return IndicatorStyles.DOT;
+  public void init(final ToolWindow window) {
+
+  }
+
+  @Override
+  public boolean shouldBeAvailable(@NotNull final Project project) {
+    return true;
+  }
+
+  @Override
+  public boolean isDoNotActivateOnStart() {
+    return false;
   }
 }