+++ src/main/java/com/chrisrm/ideaddlc/utils/MTUiUtils.java
@@ -200,7 +200,7 @@ public enum MTUiUtils {
   public static String getVersion() {
      return getPlugin()
              .map(IdeaPluginDescriptor::getVersion)
-             .orElse("4.3.3");
+             .orElse("4.4.0");
   }
 
   private static Optional<String> getPluginId() {