+++ src/main/java/io/acari/DDLC/icons/TintedColorPatcher.kt
@@ -1,10 +1,11 @@
 package io.acari.DDLC.icons
 
-import com.chrisrm.ideaddlc.MTConfig
 import com.chrisrm.ideaddlc.listeners.AccentsListener
 import com.chrisrm.ideaddlc.listeners.MTTopics
 import com.chrisrm.ideaddlc.listeners.ThemeListener
 import com.chrisrm.ideaddlc.utils.MTAccents
+import com.intellij.openapi.Disposable
+import com.intellij.openapi.application.ApplicationManager
 import com.intellij.ui.ColorUtil
 import com.intellij.util.SVGLoader
 import com.intellij.util.messages.MessageBusConnection
@@ -12,33 +13,51 @@ import io.acari.DDLC.DDLCThemeFacade
 import org.jetbrains.annotations.NonNls
 import org.w3c.dom.Element
 import java.awt.Color
+import java.net.URL
 
-class TintedColorPatcher internal constructor(connect: MessageBusConnection) : SVGLoader.SvgColorPatcher {
+class TintedColorPatcher internal constructor() : SVGLoader.SvgElementColorPatcherProvider, Disposable {
+  private val messageBusConnection: MessageBusConnection =
+      ApplicationManager.getApplication().messageBus.connect()
 
+  override fun dispose() {
+    messageBusConnection.disconnect()
+  }
   init {
-    SVGLoader.setColorPatcher(this)
+    SVGLoader.setColorPatcherProvider(this)
     val self = this
 
     // Listen for changes on the settings
-    connect.subscribe(MTTopics.ACCENTS, object : AccentsListener {
+    messageBusConnection.subscribe(MTTopics.ACCENTS, object : AccentsListener {
       override fun accentChanged(accentColor: Color) {
-        SVGLoader.setColorPatcher(null)
-        SVGLoader.setColorPatcher(self)
-        TintedColorPatcher.refreshAccentColor(accentColor)
+        SVGLoader.setColorPatcherProvider(null)
+        SVGLoader.setColorPatcherProvider(self)
+        refreshAccentColor(accentColor)
       }
     })
 
-    connect.subscribe(MTTopics.THEMES, object : ThemeListener {
+    messageBusConnection.subscribe(MTTopics.THEMES, object : ThemeListener {
       override fun themeChanged(theme: DDLCThemeFacade) {
-        SVGLoader.setColorPatcher(null)
-        SVGLoader.setColorPatcher(self)
-        TintedColorPatcher.refreshThemeColor(theme)
+        SVGLoader.setColorPatcherProvider(null)
+        SVGLoader.setColorPatcherProvider(self)
+        refreshThemeColor(theme)
       }
     })
-    refreshColors()
   }
 
-  override fun patchColors(@NonNls svg: Element) {
+  override fun forURL(url: URL?): SVGLoader.SvgElementColorPatcher? {
+    val self = this
+    return object : SVGLoader.SvgElementColorPatcher {
+      override fun patchColors(svg: Element) {
+        self.patchColors(svg)
+      }
+
+      override fun digest(): ByteArray? {
+        return null
+      }
+    }
+  }
+
+  fun patchColors(@NonNls svg: Element) {
     val tint = svg.getAttribute("tint")
     val themed = svg.getAttribute("themed")
     val themedStartAttr = svg.getAttribute("themedStart")
@@ -86,8 +105,6 @@ class TintedColorPatcher internal constructor(connect: MessageBusConnection) : S
     private var themedStart = MTAccents.CYAN.hexColor
     private var themedStop = MTAccents.TURQUOISE.hexColor
 
-    private val CONFIG = MTConfig.getInstance()
-
     internal fun refreshAccentColor(accentColor: Color) {
       TintedColorPatcher.accentColor = ColorUtil.toHex(accentColor)
     }
@@ -96,11 +113,6 @@ class TintedColorPatcher internal constructor(connect: MessageBusConnection) : S
       themedColor = ColorUtil.toHex(theme.theme.tintedIconColor)
       themedStart = theme.theme.startColor
       themedStop = theme.theme.stopColor
-    }
-
-    private fun refreshColors() {
-      accentColor = CONFIG.accentColor
-      themedColor = ColorUtil.toHex(CONFIG.selectedTheme.theme.tintedIconColor)
     }
   }
 }