+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 # 4.4.0
 
 - 2019.3 Build Support!
+- Actually supporting Heart Indicator now.
 - Up to date with Material 4.8.1
     - Better PHP Syntax Highlighting
     - More Icons!