+++ docs/CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 ---
 
+# 4.4.0
+
+- 2019.3 Build Support!
+- Actually supporting Heart Indicator now.
+- Up to date with Material 4.8.1
+    - Better PHP Syntax Highlighting
+    - More Icons!
+
 # 4.3.3
 
 - New users have to opt-in to file colors.